refactor(BcaSelect): move component to its own directory
