+++ src/components/drawers/Settings/Materials/MaterialsDrawer.tsx
@@ -8,7 +8,7 @@ import ButtonGroup from '~components/buttons/button-group'
 import DrawerTitle from '~components/titles/DrawerTitle'
 import BcaDrawer from '~components/drawers/BcaDrawer/BcaDrawer'
 import BcaTextField from '~components/input/BcaTextField'
-import BcaSelect from '~components/input/BcaSelect'
+import BcaSelect from '~/components/input/BcaSelect/BcaSelect'
 import { useGetAllCategoriesQuery } from '~redux/api/bca-backend/parametros/categoriesSlice'
 import {
     useCreateMaterialMutation,
