+++ src/components/drawers/Settings/RubroMaterial/RubroMaterialsDrawer.tsx
@@ -9,7 +9,7 @@ import DrawerTitle from '~components/titles/DrawerTitle'
 import ButtonGroup from '~components/buttons/button-group'
 import BcaTextField from '~components/input/BcaTextField'
 import { useGetAllMaterialsQuery } from '~redux/api/bca-backend/parametros/materialsSlice'
-import BcaSelect from '~components/input/BcaSelect'
+import BcaSelect from '~/components/input/BcaSelect/BcaSelect'
 import {
     useCreateRubrosMaterialMutation,
     useUpdateRubrosMaterialMutation,
