+++ src/components/drawers/Transactions/InvoiceDetailsDrawer/InvoiceDetailsDrawer.tsx
@@ -6,7 +6,7 @@ import { Stack, Typography } from '@mui/material'
 import DrawerTitle from '~components/titles/DrawerTitle'
 import BcaDrawer from '~components/drawers/BcaDrawer/BcaDrawer'
 import ButtonGroup from '~components/buttons/button-group'
-import BcaSelect from '~components/input/BcaSelect'
+import BcaSelect from '~/components/input/BcaSelect/BcaSelect'
 import { useGetAllBudgetItemsQuery } from '~redux/api/bca-backend/parametros/budgetItemSlice'
 import {
     invoiceDetailsCreateSchema,
