+++ src/pages/reportes/balance/index.tsx
@@ -6,7 +6,7 @@ import { CircularProgress, Grid2, Stack, Typography } from '@mui/material'
 
 import PageTitle from '~components/titles/PageTitle'
 import BcaDateTextField from '~/components/input/BcaDateTextField/BcaDateTextField'
-import BcaSelect from '~components/input/BcaSelect'
+import BcaSelect from '~/components/input/BcaSelect/BcaSelect'
 import EditToolbar from '~components/table/headers/toolbar'
 import { useGetAllProjectsQuery } from '~redux/api/bca-backend/parametros/projectsSlice'
 import BalanceTable from '~components/reports/BalanceTable'
