+++ src/pages/transacciones/cierre/index.tsx
@@ -10,7 +10,7 @@ import { DevTool } from '@hookform/devtools'
 import PageTitle from '~components/titles/PageTitle'
 import { useGetAllProjectsQuery } from '~redux/api/bca-backend/parametros/projectsSlice'
 import { cierreSchema, CierreTypes } from '~types/cierre'
-import BcaSelect from '~components/input/BcaSelect'
+import BcaSelect from '~/components/input/BcaSelect/BcaSelect'
 import BcaDateTextField from '~/components/input/BcaDateTextField/BcaDateTextField'
 import ConfirmationDialog from '~components/dialog/ConfirmationDialog'
 import { useCreateClosureMutation } from '~redux/api/bca-backend/transacciones/closureSlice'
